old-tests/submission/Microsoft/video/video_024.htm

<!DOCTYPE HTML>
<html>
<head>
  <title>HTML5 Media Elements: removing the 'src' attribute does not invoke the media element load algorithm.</title>
  <meta content="text/html; charset=UTF-8" http-equiv="Content-Type">
  <link rel="author" title="Microsoft" href="http://www.microsoft.com/" />
  <link rel="help" href="http://www.w3.org/TR/html5/video.html#attr-media-src" />
  <meta name="assert" content="removing the 'src' attribute does not invoke the media element load algorithm." />
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script type="text/javascript" src="../common/media.js"></script>
  <script type="text/javascript">
  
  
  var error_count = 0;
  
  var test = async_test("removing the 'src' attribute does not invoke the media element load algorithm.")
  var onerror = async_test("onerror event fires")
    
  function getTestVideoElement()
  {
      return document.getElementById("video0");
  }
  
  function do_loadstart(event)
  {
    clearTimeout(timer);
    test.step(function(){ assert_true(false, "'loadstart' event fired - media loading!")});
    test.done();
  }
  
  function do_error(event)
  {
      error_count ++;
      
      if (error_count == 1)
      {
          getTestVideoElement().addEventListener("loadstart", do_loadstart, false);
          getTestVideoElement().removeAttribute("src");
          onerror.step(function(){assert_true(true)});
          onerror.done();
      }
  }
  
  function timeout()
  {
    onerror.step(function(){assert_true(false)});
    onerror.done();
    test.step(function(){assert_true(true)});
    test.done();
    
  }
  
  var timer = setTimeout(timeout, 4000);
  </script>
</head>
<body>
  <div id=log></div>
  <video id="video0" src="broken-uri" autoplay onerror="do_error(event);">
  <source src="broken-uri-2" type="video/mp4">
  <source src="broken-uri-3" type="video/ogg">
  Your browser does not support media elements.
  </video>
</body>
</html>
